HF-1.8: Fix bag in publicity add loger in Global and refactoring code

+++ OnlineStore_Epam2018/SA.OnlineStore.Bussines/Service/Implementation/ProductService.cs
@@ -20,16 +20,26 @@
 
         public void DeleteProductByProductId(int Id) 
         {
-            _productRepository.Delete(Id);
+            if (Id > 0)
+            {
+                _productRepository.Delete(Id);
+            }
         }
 
         public void EditProduct(Product model)
         {
-            _productRepository.Update(model);
+            if (model != null)
+            {
+                _productRepository.Update(model);
+            }
         }
 
         public Product GetProduct(int Id)
         {
+            if (Id < 1)
+            {
+                return null;
+            }
             return _productRepository.GetById(Id);
         }
 
@@ -40,12 +50,19 @@
 
         public IEnumerable<Product> GetProductLIstByCategory(int category)
         {
+            if (category < 0)
+            {
+                return null;
+            }
             return _productRepository.GetAll().Where(x => x.Category.CategoryId == category);
         }
 
         public void SaveProduct(Product model)
         {
-            _productRepository.Create(model);
+            if (model != null)
+            {
+                _productRepository.Create(model);
+            }
         }
     }
 }
